requirements.txt
amqp==2.5.2
certifi==2019.11.28
chardet==3.0.4
contextlib2==0.5.5
dnspython==1.16.0
eventlet==0.25.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.26.0
grpcio-tools==1.26.0
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
idna==2.8
importlib-metadata==1.3.0
iris-model==0.1.0
Jinja2==2.10.3
joblib==0.14.1
kombu==4.6.7
MarkupSafe==1.1.1
mock==3.0.5
monotonic==1.5
more-itertools==8.0.2
nameko==2.12.0
nameko-grpc==1.0.1
numpy==1.18.0
path==13.1.0
path.py==12.4.0
protobuf==3.11.2
PyYAML==5.2
requests==2.22.0
schema==0.7.0
scikit-learn==0.21.0
scipy==1.4.1
six==1.13.0
urllib3==1.25.7
vine==1.3.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
